requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1
google-generativeai==0.8.6
chromadb==0.4.22
pypdf2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openai==1.10.0
httpx==0.26.0
python-dotenv==1.0.0
numpy==1.26.3
sentence-transformers==2.3.1
langchain==0.1.4
langchain-community==0.0.17
langchain-google-genai==0.0.5
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1
